refactor: 라우팅이 아닌 폴더를 private folter로 변경
https://nextjs.org/docs/app/getting-started/project-structure#private-folders
