+++ packages/web/src/app/[lng]/(auth)/memos/setting/page.tsx
@@ -6,7 +6,7 @@ import { Header, HydrationBoundaryWrapper } from '@src/components';
 import { LanguageParams } from '@src/modules/i18n';
 import { getSupabaseClient } from '@src/modules/supabase/util.server';
 
-import { Setting, SettingHeader } from './component';
+import { Setting, SettingHeader } from './_components';
 
 export default async function Page({ params: { lng } }: LanguageParams) {
   const supabaseClient = getSupabaseClient();
