+++ packages/web/src/app/[lng]/(no-auth)/introduce/_components/AdditionalFeatures/index.tsx
