+++ packages/web/src/app/[lng]/(no-auth)/introduce/_components/Features/index.tsx
