+++ packages/web/src/app/[lng]/(no-auth)/introduce/_components/Hero/index.tsx
