+++ packages/web/src/app/[lng]/(no-auth)/introduce/_components/ImageSlider/index.tsx
