+++ packages/web/src/app/[lng]/(no-auth)/introduce/_components/QuestionAndAnswer/index.tsx
