+++ packages/web/src/app/[lng]/(no-auth)/login/_components/LoginSection/ExtensionVersion.tsx
