+++ packages/web/src/app/[lng]/(no-auth)/login/_components/PersonalInformationInfo/index.tsx
