+++ packages/web/src/app/[lng]/(no-auth)/update/_components/UpdateList.tsx
