+++ packages/web/src/app/[lng]/(no-auth)/update/_components/index.ts
