+++ packages/web/src/app/_components/ExtensionDialog/useExtensionDialogInfo.ts
